refactor: move inmemory file-writer to fake namespace
