+++ packages/skott/test/unit/runner.spec.ts
@@ -1,7 +1,7 @@
 import { describe, expect, test } from "vitest";
 
 import { InMemoryFileReader } from "../../src/filesystem/fake/file-reader.js";
-import { InMemoryFileWriter } from "../../src/filesystem/file-writer.js";
+import { InMemoryFileWriter } from "../../src/filesystem/fake/file-writer.js";
 import { FakeLogger } from "../../src/logger.js";
 import { ModuleWalkerSelector } from "../../src/modules/walkers/common.js";
 import { defaultConfig, Skott } from "../../src/skott.js";
