+++ packages/skott/test/unit/shared.ts
@@ -1,7 +1,7 @@
 import * as memfs from "memfs";
 
 import { InMemoryFileReader } from "../../src/filesystem/fake/file-reader.js";
-import { InMemoryFileWriter } from "../../src/filesystem/file-writer.js";
+import { InMemoryFileWriter } from "../../src/filesystem/fake/file-writer.js";
 import type { SkottNode } from "../../src/graph/node.js";
 import { FakeLogger } from "../../src/logger.js";
 import { kExpectedModuleExtensions } from "../../src/modules/resolvers/base-resolver.js";
